ICM20948_ReadRegisters size uint16 ->uint8

+++ Inc/icm20948.h
@@ -542,7 +542,7 @@ HAL_StatusTypeDef ICM20948_ReadUserBank(uint8_t* data);
 HAL_StatusTypeDef ICM20948_ChangeUserBank(user_bank ubank);
 
 HAL_StatusTypeDef ICM20948_ReadRegister(const reg_R* regi, uint8_t* data);
-HAL_StatusTypeDef ICM20948_ReadRegisters(const reg_R* regi, uint8_t* data, uint16_t size);
+HAL_StatusTypeDef ICM20948_ReadRegisters(const reg_R* regi, uint8_t* data, uint8_t size);
 HAL_StatusTypeDef ICM20948_WriteRegister(const reg_RW* regi, uint8_t data);
 HAL_StatusTypeDef ICM20948_WriteRegisterEnables(const reg_RW* regi, uint8_t data);
 HAL_StatusTypeDef ICM20948_WriteRegisterDisables(const reg_RW* regi, uint8_t data);
